requirements.txt
fastapi==0.95.2
uvicorn==0.27.1
websockets==12.0
python-dotenv==1.0.0
openai==1.12.0
pinecone-client==3.0.2
langchain==0.1.9
langchain-openai==0.0.5
langchain-pinecone==0.0.3
jinja2==3.1.4
pydantic>=1.10.0,<2.0.0
